chore: be more surgical about what is moved in utils
